+++ tests/unit/test_dataset_records.py
@@ -99,6 +99,33 @@ class TestDatasetRecords(TestCase):
         self.assertEqual(response, self._response_json)
         self.assertEqual(snoop["payload"], TestDatasetRecords.stringify(updates, False))
 
+    @responses.activate
+    def test_delete(self):
+        def create_callback(request, snoop):
+            snoop["payload"] = list(request.body)
+            return 200, {}, simplejson.dumps(self._response_json)
+
+        responses.add(responses.GET, self._dataset_url, json={})
+        dataset = self.unify.datasets.by_resource_id(self._dataset_id)
+
+        records_url = f"{self._dataset_url}:updateRecords"
+        deletes = TestDatasetRecords.records_to_deletes(self._records_json)
+        snoop = {}
+        responses.add_callback(
+            responses.POST, records_url, partial(create_callback, snoop=snoop)
+        )
+
+        response = dataset.delete_records(self._records_json, "attribute1")
+        self.assertEqual(response, self._response_json)
+        self.assertEqual(snoop["payload"], TestDatasetRecords.stringify(deletes, False))
+
+    @staticmethod
+    def records_to_deletes(records):
+        return [
+            {"action": "DELETE", "recordId": i}
+            for i, record in enumerate(records, start=1)
+        ]
+
     @staticmethod
     def records_to_updates(records):
         return [